requirements.txt
selene==1.0.2
selenium==3.141.0
PyHamcrest==2.0.2
pytest==6.2.3